Change VoteMap type, fix helper functions
The VoteMap mapping is now bytes32 => uint instead of bytes32 => bytes32, and
thus helper functions, except getCommitHash, returns uints. The reason this type
change was made is that when uint variables implicitly got cast to bytes32, which
happened during mapping, they would end up on the left side of the hexadecimal
number instead of the right side, causing the number to be extraordinarily large.

Helper functions were fixed according to the change above, and to protect the
integrity of commitHashs, a seperate helper function to get it from mapping was
implemented.

Lastly, to increase the ease of testing, a helper function that produces the
exact sha3 hash in Javascript that Solidity produces when hashing an address,
a uint and a string was implemented and named createIndexHash. This is needed
since a hash of an address, a uint and a string is used as index for the VoteMap
mapping.