Fix getInsertPoint failng for in-place updates
The getInsertPointForNumTokens function fails to provide correct insert
points in a number of cases involving in-place updates.

For in-place updates where the new value is greater than or equal to
the old value, the function will return, incorrectly, the insert point
as being the same pollID as the poll being updated.

For in-place updates where the node being updated's token value is
zero, the function will always compute, incorrectly, that the correct
insert point is at the end of the list.

* Change the getInsertPoint function signature to accept the ID of the
  node being inserted or updated
* Take a node as being a correct insert point if the number of tokens it
  contains is less than or equal to (rather than strictly less than) the
  number of tokens being inserted
* If the computed insert point is the same as the provided pollID, the
  operation is an in-place update, and the previous node of the node
  being updated is the correct insert point
* Stop iterating backwards through the list when the NULL_NODE is
  reached, rather than when a node with zero tokens is reached
* Update usages of getInsertPointForNumTokens in the tests to use its
  new function signature
* getInsertPointForNumTokens: add tests for in-place list updates
* listSanity: add tests for 4 cases of inserting incorrect prevPollIDs,
  and assert that such attempted insertions will revert
* Add comments