Update package.json
It is currently difficult to depend on this project through importing them into solidity.

I don't understand how someone thought it was a good idea to name the repo differently for `npm` vs. `ethpm` but since one of those package-managers is on an immutable _blockchain_, I figure it is less disruptive to change this one.

With the proposed change I would be able to write my Solidity code compatible with importing it using `npm` or `ethpm`.