Add getter functions
Previously, it was difficult to determine whether a voter had
committed and/or revealed in a poll. One would have to collect all
`_VoteCommitted` and/or `_VoteRevealed` logs to build up the context of
each poll. This commit adds 2 constant getter functions enabling access to the
`didCommit` and `didReveal` mappings of a poll.