inline requestVotingRights into commitVote
previously, commitVote required msg.sender to have a voteTokenBalance greater than
or equal to the number of tokens the msg.sender wanted to commit. this
required an additional transaction to be sent before a user could interact
with the PLCRVoting contract (3 txns: token.approve,
plcr.requestVotingRights, plcr.commitVote). this commit inlines
requestVotingRights so that a user can call commitVote directly after
approval (2 txns).

* change visibility of requestVotingRights from external -> public so it
can be invoked internally
* invoke requestVotingRights from commitVote if voteTokenBalance is less
than the requested _numTokens for commitVote
* observe 1 failing commitVote test