Make the PLCRFactory more trustless
Previously, the PLCRVoting factory would take in the address of a
deployed PLCRVoting contract for canonization. But trusting this
requires the user verify the code at the address input in the
constructor was really a PLCRVoting contract. Instead, we can just
deploy a new PLCRVoting in the factory constructor, such that the user
only needs to audit one contract.

* Canonize a PLCRVoting contract by deploying one in the factory
  constructor
* Remove migrations that deploy PLCRVoting